requirements-dev.txt
attrs==22.1.0
aws-cdk-lib==2.41.0
black==22.8.0
cattrs==22.1.0
cfgv==3.3.1
click==8.1.3
constructs==10.1.104
distlib==0.3.6
exceptiongroup==1.0.0rc9
filelock==3.8.0
identify==2.5.5
iniconfig==1.1.1
jsii==1.67.0
mypy-extensions==0.4.3
nodeenv==1.7.0
packaging==21.3
pathspec==0.10.1
platformdirs==2.5.2
pluggy==1.0.0
pre-commit==2.20.0
publication==0.0.3
py==1.11.0
pyparsing==3.0.9
pytest==6.2.5
python-dateutil==2.8.2
PyYAML==6.0
six==1.16.0
toml==0.10.2
tomli==2.0.1
typeguard==2.13.3
typing_extensions==4.3.0
virtualenv==20.16.5
